move checks.py along with check-specific errors to ext module
