+++ nextcord/ext/checks/__init__.py
@@ -1,0 +1,13 @@
+"""
+nextcord.ext.commands
+~~~~~~~~~~~~~~~~~~~~~
+
+An extension module to handle application command checks.
+
+:copyright: (c) 2015-present Rapptz
+:copyright: (c) 2021-present tag-epic
+:license: MIT, see LICENSE for more details.
+"""
+
+from .core import *
+from .errors import *
